requirements.txt
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
Flask-Session==0.4.0